All deletes or removes now prompt user

+++ wwwsrc/src/components/VaultKeeps.vue
@@ -20,6 +20,7 @@
 
 
 <script>
+import NotificationService from "../NotificationService";
 export default {
   name: "VaultKeeps",
   props: ["keepProp"],
@@ -33,18 +34,20 @@ export default {
     }
   },
   methods: {
-    removeKeep() {
-      this.$store.dispatch("removeFromVault", {
-        keepId: this.keepProp.id,
-        vaultId: this.vault.id
-      });
-      this.$store.dispatch("updateKeepsTotal", {
-        id: this.keepProp.id,
-        views: this.keepProp.views,
-        name: this.keepProp.name,
-        img: this.keepProp.img,
-        keeps: (this.keepProp.keeps -= 1)
-      });
+    async removeKeep() {
+      if (await NotificationService.remove()) {
+        this.$store.dispatch("removeFromVault", {
+          keepId: this.keepProp.id,
+          vaultId: this.vault.id
+        });
+        this.$store.dispatch("updateKeepsTotal", {
+          id: this.keepProp.id,
+          views: this.keepProp.views,
+          name: this.keepProp.name,
+          img: this.keepProp.img,
+          keeps: (this.keepProp.keeps -= 1)
+        });
+      }
     },
     viewKeep() {
       this.$store.dispatch("getKeepById", this.keepProp.id);
